Start to write script to "mount" roam directories
(sorry for poor syntax, I'm tired so I'll be verbose and probably redundant)

Doom Emacs's Org-roam graph generation requires a specific directory to be used for graph generation. Since there's a lot of info that can be represented using Org-roam graph capabilities, very often I want to generate different graphs for different topics. However, this is not quite intuitive using Org-roam's design.

First you have to define a "roam directory" or something like that so Org-roam can use that directory to place the roam files and then generate the graph. At least that's how I understand it. Anyway, every time I want to create a graph I have to place the files in this "roam directory" and only then generate the graph. If I want to do another completely different without losing the work of other projects I have to back up whatever it was before in that "roam directory", then create the files in this directory and finally generate the graph.

The ~/bin/roam Bash scripts takes care of that process. Maybe in the future Org-roam will solve this issue but for now, I'll be happy to hack it this way.

The "roam directory" that I use by default (defined in the ~/.doom.d/config.el") is ~/org/auxRoam"; anything I want to generate a graph from, has to be placed inside this auxRoam directory, create the graph and then backup this roam files in their correspondent place.

Actually, the first solution I tried was to manually change the "org-roam-directory" in the Doom Emacs configuration itself, but it quickly became an annoyance since I had to change it every time I wanted to graph different ideas. Hence, this script.