Disable spell config & add documentation.
1) Disable spell config:
explained in previous commit; now it's just the org-file part.

2) Add documentation for why _not_ activate a variable related to
fontification in org mode --- the fontification only takes place in the
buffer but not to export. And I want the opposite; I don't mind the in-buffer
fontification as long as export fontification takes place, but that is not
the case; I still have to put =/= around words for them to be fontified,
even if they are inside a =quote= block.