feat(emacs): encrypt agenda-files.txt file
Now it is an encrypted version of agenda files
that is being tracked.

A local unencrypted version is being used
as agenda-files actual file, but an
encrypted version of it is in ~/FilenSync/ directory.

The location of 'agenda-files.txt' is defined
in '~/.doom.d/config.org' and its value is
'~/agenda-files.txt'.