feat(bin): add 'orgpdfpreview'  !!!
This is a small utility that let's you
write temporary orgmode files and visualize
the exported pdf without using the emacs
default export page which is slower,
specially considering when you add small changes
to the orgmode file and just want to visualize
those small changes without the overhead of
compiling the entire document every time.

Maybe in the future will implement for html export
too.